update service 10/12/22

+++ back-end/AssetManagementTeam6.API/Services/Implements/AssignmentService.cs
@@ -25,9 +25,7 @@ namespace AssetManagementTeam6.API.Services.Implements
         public async Task<IEnumerable<GetAssetResponse>> CheckAvailableAsset()
         {
             var assetIds = new List<int>();
-
-            var assignments = await _assignmentRepository.GetListAsync();
-
+            var assignments = await _assignmentRepository.GetListAsync(x => x.State != AssignmentStateEnum.Deleted);
             var assignedAssets = assignments.Select(assignment => assignment.Asset.Id);
 
             foreach (var assetCode in assignedAssets)
@@ -48,9 +46,29 @@ namespace AssetManagementTeam6.API.Services.Implements
 
             var assignedBy = await _userRepository.GetOneAsync(u => u.Id == createRequest.AssignedById);
 
-            if (asset == null || assignedTo == null || assignedBy == null)
+            if(asset == null)
             {
-                return null;
+                throw new Exception("Asset is not exist");
+            }
+
+            if (asset.State == AssetStateEnum.Assigned)
+            {
+                throw new Exception("Asset is assigned for someone else");
+            }
+
+            if (asset.State != AssetStateEnum.Available)
+            {
+                throw new Exception("Asset is not available");
+            }
+
+            if(assignedTo == null)
+            {
+                throw new Exception("Assignee is not exist");
+            }
+
+            if(assignedBy == null)
+            {
+                throw new Exception("Assigner is not exist");
             }
 
             var newAssignment = new Assignment
@@ -78,59 +96,55 @@ namespace AssetManagementTeam6.API.Services.Implements
 
         public async Task<IEnumerable<GetAssignmentResponse>> GetAllAsync()
         {
-            var assignments = await _assignmentRepository.GetListAsync();
+            var assignments = await _assignmentRepository.GetListAsync(x => x.State != AssignmentStateEnum.Deleted);
 
             return assignments.Select(ass => new GetAssignmentResponse(ass)).ToList();
         }
 
-        public async Task<Assignment> GetAssignmentByAssignedAsset(int assetId)
+        public async Task<bool> GetAssignmentByAssignedAsset(int assetId)
         {
-            var result = await _assignmentRepository.GetOneAsync(a => a.AssetId == assetId);
+            var result = await _assignmentRepository.GetListAsync(x => x.AssetId == assetId && x.State != AssignmentStateEnum.Deleted);
 
-            if (result == null)
+            if (result.Count() == 0 || result == null)
             {
-                return null!;
+                return false;
             }
 
-            return result;
+            return true;
         }
 
-        public async Task<Assignment> GetAssignmentByAssignedUser(int assignedUserId)
+        public async Task<bool> GetAssignmentByAssignedUser(int assignedUserId)
         {
-            var result = await _assignmentRepository.GetOneAsync(a => a.AssignedToId == assignedUserId);
-            if (result == null)
-            {
-                return null!;
-            }
+            var result = await _assignmentRepository.GetListAsync(a => a.AssignedToId == assignedUserId && a.State != AssignmentStateEnum.Deleted);
 
-            return result;
+            if (result.Count() == 0 || result == null)
+            {
+                return false;
+            }           
+
+            return true;
         }
 
         public async Task<IEnumerable<GetAssignmentResponse>> GetListAssignmentByUserLoggedIn(int id)
         {
-            // TODO: DateTime
-            // var assignments = await _assignmentRepository.GetListAsync(ass => ass.AssignedToId == id && ass.AssignedDate <= DateTime.UtcNow);
-
-            var assignments = await _assignmentRepository.GetListAsync(ass => ass.AssignedToId == id);
+            var assignments = await _assignmentRepository.GetListAsync(ass => ass.AssignedToId == id 
+                                                                    && ass.State != AssignmentStateEnum.Deleted
+                                                                    && ass.State != AssignmentStateEnum.Declined
+                                                                    && ass.AssignedDate <= DateTime.UtcNow);
 
             return assignments.Select(ass => new GetAssignmentResponse(ass)).ToList();
         }
 
         public async Task<Assignment?> GetAssignmentById(int id)
         {
-            return await _assignmentRepository.GetOneAsync(a => a.Id == id);
+            return await _assignmentRepository.GetOneAsync(x => x.Id == id && x.State != AssignmentStateEnum.Deleted);
 
-        }
-
-        public async Task<Assignment?> GetAssignmentByAssetId(int AssetId)
-        {
-            return await _assignmentRepository.GetOneAsync(a => a.AssetId == AssetId);
         }
 
         public async Task<Pagination<GetAssignmentResponse?>> GetPagination(PaginationQueryModel queryModel)
         {
             // get all list
-            var assignments = await _assignmentRepository.GetListAsync();
+            var assignments = await _assignmentRepository.GetListAsync(x => x.State != AssignmentStateEnum.Deleted);
            
             // search assignment by assetcode or assetname or assignee
             var nameToQuery = "";
@@ -231,11 +245,26 @@ namespace AssetManagementTeam6.API.Services.Implements
 
         public async Task<GetAssignmentResponse> AcceptedAssignment(int id)
         {
-            var updatedAssignment = await _assignmentRepository.GetOneAsync(x => x.Id == id);
+            var updatedAssignment = await _assignmentRepository.GetOneAsync(x => x.Id == id && x.State != AssignmentStateEnum.Deleted);
 
             if(updatedAssignment == null)
             {
-                return null!;
+                throw new Exception("Assignment is not exist");
+            }
+
+            if(updatedAssignment.State != AssignmentStateEnum.WaitingForAcceptance)
+            {
+                throw new Exception("Assignment state is not waiting for acceptance");
+            }
+
+            if(updatedAssignment.Asset == null)
+            {
+                throw new Exception("Asset is not exist");
+            }
+
+            if(updatedAssignment.Asset.State != AssetStateEnum.Available)
+            {
+                throw new Exception("Asset state not available");
             }
 
             updatedAssignment.Asset.State = AssetStateEnum.Assigned;
@@ -249,11 +278,26 @@ namespace AssetManagementTeam6.API.Services.Implements
 
         public async Task<GetAssignmentResponse> DeclinedAssignment(int id)
         {
-            var updatedAssignment = await _assignmentRepository.GetOneAsync(x => x.Id == id);
+            var updatedAssignment = await _assignmentRepository.GetOneAsync(x => x.Id == id && x.State != AssignmentStateEnum.Deleted);
 
             if (updatedAssignment == null)
             {
-                return null!;
+                throw new Exception("Assignment is not exist");
+            }
+
+            if (updatedAssignment.State != AssignmentStateEnum.WaitingForAcceptance)
+            {
+                throw new Exception("Assignment state is not waiting for acceptance");
+            }
+
+            if (updatedAssignment.Asset == null)
+            {
+                throw new Exception("Asset is not exist");
+            }
+
+            if (updatedAssignment.Asset.State != AssetStateEnum.Available)
+            {
+                throw new Exception("Asset state not available");
             }
 
             var assetState = updatedAssignment.Asset.State;
@@ -272,23 +316,42 @@ namespace AssetManagementTeam6.API.Services.Implements
             return new GetAssignmentResponse(result!);
         }
 
-        //TODO
         public async Task<GetAssignmentResponse?> Update(int id, AssignmentRequest? updatedRequest)
         {
-            var updatedAssignment = await _assignmentRepository.GetOneAsync(x => x.Id == id);
+            var updatedAssignment = await _assignmentRepository.GetOneAsync(x => x.Id == id && x.State != AssignmentStateEnum.Deleted);
             var assignedTo = await _userRepository.GetOneAsync(x => x.Id == updatedRequest!.AssignedToId);
             var assignedBy = await _userRepository.GetOneAsync(x => x.Id == updatedRequest!.AssignedById);
-            var updateAsset = await _assetRepository.GetOneAsync(x => x.Id == updatedRequest!.AssetId );
+            var updatedAsset = await _assetRepository.GetOneAsync(x => x.Id == updatedRequest!.AssetId );
 
-            if(updatedAssignment == null || assignedTo == null || updateAsset == null || updateAsset == null)
+            if(updatedAssignment == null)
             {
-                return null;
+                throw new Exception("Assignment is not exist");
+            }
+
+            if(updatedAssignment.State != AssignmentStateEnum.WaitingForAcceptance)
+            {
+                throw new Exception("Invalid state");
+            }
+
+            if(updatedAsset == null)
+            {
+                throw new Exception("Asset is not exist");
+            }
+
+            if(updatedAsset.State != AssetStateEnum.Available)
+            {
+                throw new Exception("Asset not available");
+            }
+
+            if(assignedTo == null)
+            {
+                throw new Exception("Asignee is not exist");
             }
             
             updatedAssignment.Note = updatedRequest?.Note ?? "";
             updatedAssignment.AssignedDate = updatedRequest!.AssignedDate;
             updatedAssignment.AssetId = updatedRequest.AssetId;
-            updatedAssignment.Asset = updateAsset;
+            updatedAssignment.Asset = updatedAsset;
             updatedAssignment.AssignedToId = updatedRequest.AssignedToId;
             updatedAssignment.AssignedTo = assignedTo;
             updatedAssignment.AssignedById = updatedRequest.AssignedById;
@@ -303,11 +366,16 @@ namespace AssetManagementTeam6.API.Services.Implements
 
         public async Task<bool> Delete(int id)
         {
-            var deletedAssignment = await _assignmentRepository.GetOneAsync(x => x.Id == id);
+            var deletedAssignment = await _assignmentRepository.GetOneAsync(x => x.Id == id && x.State != AssignmentStateEnum.Deleted);
 
             if (deletedAssignment == null)
             {
-                return false;
+                throw new Exception("Assignment is not exist");
+            }
+
+            if(deletedAssignment.State != AssignmentStateEnum.WaitingForAcceptance || deletedAssignment.State != AssignmentStateEnum.Declined)
+            {
+                throw new Exception("Invalid state");
             }
 
             return await _assignmentRepository.Delete(deletedAssignment);
